+++ lib/cocoapods/installer/target_installer/aggregate_target_installer.rb
@@ -32,15 +32,14 @@ module Pod
       # @return [void]
       #
       def create_xcconfig_file
-        path = library.xcconfig_path
-        UI.message "- Generating xcconfig file at #{UI.path(path)}" do
-          gen = Generator::XCConfig::AggregateXCConfig.new(library)
-          gen.save_as(path)
-          library.xcconfig = gen.xcconfig
-          xcconfig_file_ref = add_file_to_support_group(path)
-
-          target.build_configurations.each do |c|
-            c.base_configuration_reference = xcconfig_file_ref
+        target.build_configurations.each do |configuration|
+          path = library.xcconfig_path(configuration.name)
+          UI.message "- Generating #{configuration.name} xcconfig file at #{UI.path(path)}" do
+            gen = Generator::XCConfig::AggregateXCConfig.new(library, configuration.name)
+            gen.save_as(path)
+            library.xcconfigs[configuration.name] = gen.xcconfig
+            xcconfig_file_ref = add_file_to_support_group(path)
+            configuration.base_configuration_reference = xcconfig_file_ref
           end
         end
       end
@@ -51,7 +50,7 @@ module Pod
       def create_target_environment_header
         path = library.target_environment_header_path
         UI.message "- Generating target environment header at #{UI.path(path)}" do
-          generator = Generator::TargetEnvironmentHeader.new(library.pod_targets.map { |l| l.specs }.flatten)
+          generator = Generator::TargetEnvironmentHeader.new(library.specs_by_build_configuration)
           generator.save_as(path)
           add_file_to_support_group(path)
         end