+++ CHANGELOG.md
@@ -6,6 +6,18 @@ To install or update CocoaPods see this [guide](http://docs.cocoapods.org/guides
 
 ##### Enhancements
 
+* Add support to specify dependencies per build configuration.
+
+      pod 'PonyDebugger', :configurations => ['Debug']
+
+  Currently configurations can only be specified per single Pod.  
+  [Joachim Bengtsson](https://github.com/nevyn)
+  [Eloy Durán](https://github.com/alloy)
+  [Fabio Pelosin][irrationalfab]
+  [#1791](https://github.com/CocoaPods/CocoaPods/pull/1791)
+  [#1668](https://github.com/CocoaPods/CocoaPods/pull/1668)
+  [#731](https://github.com/CocoaPods/CocoaPods/pull/731)
+
 * Added hooks for plugins. Currently only the installer hook is supported.
   A plugin can register itself to be activated after the installation with the
   following syntax:
@@ -26,6 +38,14 @@ To install or update CocoaPods see this [guide](http://docs.cocoapods.org/guides
 * Use gem CLIntegracon for the integration tests.  
   [Marius Rackwitz][mrackwitz]
   [#2371](https://github.com/CocoaPods/CocoaPods/issues/2371)
+
+* Include configurations a user explicitly specifies in their Podfile when the
+  `--no-integrate` option is specified.  
+  [Eloy Durán](https://github.com/alloy)
+
+* Properly quote the `-isystem` values in the xcconfig files.  
+  [Eloy Durán](https://github.com/alloy)
+
 
 ##### Bug Fixes